refactor: move readDescription function to cmd_description
